reuse AccelSwitchType for cstone domain
