+++ sph/include/sph/hydro_std/density.hpp
@@ -83,7 +83,7 @@ void computeDensityImpl(size_t startIndex, size_t endIndex, size_t ngmax, Datase
 template<class T, class Dataset>
 void computeDensity(size_t startIndex, size_t endIndex, size_t ngmax, Dataset& d, const cstone::Box<T>& box)
 {
-    if constexpr (sphexa::HaveGpu<typename Dataset::AcceleratorType>{})
+    if constexpr (cstone::HaveGpu<typename Dataset::AcceleratorType>{})
     {
         cuda::computeDensity(startIndex, endIndex, ngmax, d, box);
     }
