+++ sph/include/sph/hydro_std/momentum_energy.hpp
@@ -128,7 +128,7 @@ void computeMomentumEnergySTDImpl(size_t startIndex, size_t endIndex, size_t ngm
 template<class T, class Dataset>
 void computeMomentumEnergySTD(size_t startIndex, size_t endIndex, size_t ngmax, Dataset& d, const cstone::Box<T>& box)
 {
-    if constexpr (sphexa::HaveGpu<typename Dataset::AcceleratorType>{})
+    if constexpr (cstone::HaveGpu<typename Dataset::AcceleratorType>{})
     {
         cuda::computeMomentumEnergySTD(startIndex, endIndex, ngmax, d, box);
     }
