+++ sph/include/sph/hydro_ve/iad_divv_curlv.hpp
@@ -128,7 +128,7 @@ void computeIadDivvCurlvImpl(size_t startIndex, size_t endIndex, size_t ngmax, D
 template<class T, class Dataset>
 void computeIadDivvCurlv(size_t startIndex, size_t endIndex, size_t ngmax, Dataset& d, const cstone::Box<T>& box)
 {
-    if constexpr (sphexa::HaveGpu<typename Dataset::AcceleratorType>{})
+    if constexpr (cstone::HaveGpu<typename Dataset::AcceleratorType>{})
     {
         cuda::computeIadDivvCurlv(startIndex, endIndex, ngmax, d, box);
     }
