+++ sph/include/sph/hydro_ve/ve_def_gradh.hpp
@@ -81,7 +81,7 @@ void computeVeDefGradhImpl(size_t startIndex, size_t endIndex, size_t ngmax, Dat
 template<typename T, class Dataset>
 void computeVeDefGradh(size_t startIndex, size_t endIndex, size_t ngmax, Dataset& d, const cstone::Box<T>& box)
 {
-    if constexpr (sphexa::HaveGpu<typename Dataset::AcceleratorType>{})
+    if constexpr (cstone::HaveGpu<typename Dataset::AcceleratorType>{})
     {
         cuda::computeVeDefGradh(startIndex, endIndex, ngmax, d, box);
     }
